Abstraction in OOP
 all the attributes and methods of the class can be accessed through
the object. It is possible in regular classes; whereas, in the case of an abstract class, an object cannot be
instantiated. This is why it is helpful to hide information that need not be shared with external users.
Abstraction is a method of information protection in Python or any other OOP language. We will
now look at encapsulation and more details on how information can be protected in a class.